+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `googleAddress` to typed Lookup Elements
+
 ## [6.1.0] - 2024-04-30
 
 ## Added